+++ src/TrackerManagement/TrackerManager.h
@@ -146,7 +146,7 @@ public:
 
 	void controlTimeline(bool stop);
 
-	void timelineValueChange(int newValue);
+	void timelineValueChange(float newValue);
 
 	FrameData getRecCurrFrameData();
